+++ src/OctoshiftCLI.Tests/bbs2gh/BbsApiFactoryTests.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using System.Net.Http;
 using FluentAssertions;
 using Moq;
-using OctoshiftCLI.BbsToGithub;
+using OctoshiftCLI.BbsToGithub.Factories;
 using OctoshiftCLI.Services;
 using Xunit;
