+++ src/OctoshiftCLI.Tests/bbs2gh/Commands/GenerateScriptCommandTests.cs
@@ -1,8 +1,8 @@
 using System;
 using FluentAssertions;
 using Moq;
-using OctoshiftCLI.BbsToGithub;
 using OctoshiftCLI.BbsToGithub.Commands;
+using OctoshiftCLI.BbsToGithub.Factories;
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Services;
 using Xunit;
