+++ src/OctoshiftCLI.Tests/bbs2gh/Commands/MigrateRepoCommandTests.cs
@@ -1,8 +1,8 @@
 using System;
 using FluentAssertions;
 using Moq;
-using OctoshiftCLI.BbsToGithub;
 using OctoshiftCLI.BbsToGithub.Commands;
+using OctoshiftCLI.BbsToGithub.Factories;
 using OctoshiftCLI.Services;
 using Xunit;
