+++ src/bbs2gh/Commands/GenerateScriptCommand.cs
@@ -2,6 +2,7 @@ using System;
 using System.CommandLine;
 using System.IO;
 using Microsoft.Extensions.DependencyInjection;
+using OctoshiftCLI.BbsToGithub.Factories;
 using OctoshiftCLI.BbsToGithub.Handlers;
 using OctoshiftCLI.Commands;
 using OctoshiftCLI.Contracts;
