+++ src/bbs2gh/Factories/BlobServiceClientFactory.cs
@@ -1,6 +1,6 @@
 using Azure.Storage.Blobs;
 
-namespace OctoshiftCLI.BbsToGithub;
+namespace OctoshiftCLI.BbsToGithub.Factories;
 
 public class BlobServiceClientFactory : IBlobServiceClientFactory
 {
