--- src/bbs2gh/IAzureApiFactory.cs
@@ -1,10 +1,0 @@
-using OctoshiftCLI.Services;
-
-namespace OctoshiftCLI.BbsToGithub
-{
-    public interface IAzureApiFactory
-    {
-        AzureApi Create(string azureStorageConnectionString);
-        AzureApi CreateClientNoSsl(string azureStorageConnectionString);
-    }
-}
